landinghelp_server_backend/requirements/dev.txt
-r base.txt
pytest
pytest-django
ipython
